requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
